Move CNF from SHR to Tokens, move to public
