+++ src/Microsoft.IdentityModel.Tokens/ConfirmationClaimTypes.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.IdentityModel.Protocols.SignedHttpRequest
+namespace Microsoft.IdentityModel.Tokens
 {
     /// <summary>
     /// Confirmation Claim ("cnf") related constants
@@ -14,27 +14,27 @@ namespace Microsoft.IdentityModel.Protocols.SignedHttpRequest
         /// <summary>
         /// https://datatracker.ietf.org/doc/html/rfc7800#section-6.1.1
         /// </summary>
-        public const string Cnf = Tokens.ConfirmationClaimTypes.Cnf;
+        public const string Cnf = "cnf";
 
         /// <summary>
         /// https://datatracker.ietf.org/doc/html/rfc7800#section-6.2.2
         /// </summary>
-        public const string Jwk = Tokens.ConfirmationClaimTypes.Jwk;
+        public const string Jwk = "jwk";
 
         /// <summary>
         /// https://datatracker.ietf.org/doc/html/rfc7800#section-6.2.2
         /// </summary>
-        public const string Jwe = Tokens.ConfirmationClaimTypes.Jwe;
+        public const string Jwe = "jwe";
 
         /// <summary>
         /// https://datatracker.ietf.org/doc/html/rfc7800#section-6.2.2
         /// </summary>
-        public const string Jku = Tokens.ConfirmationClaimTypes.Jku;
+        public const string Jku = "jku";
 
         /// <summary>
         /// https://datatracker.ietf.org/doc/html/rfc7800#section-6.2.2
         /// </summary>
-        public const string Kid = Tokens.ConfirmationClaimTypes.Kid;
+        public const string Kid = "kid";
     }
 
     internal static class ConfirmationClaimTypesUtf8Bytes
@@ -45,5 +45,4 @@ namespace Microsoft.IdentityModel.Protocols.SignedHttpRequest
         public static ReadOnlySpan<byte> Jku => "jku"u8;
         public static ReadOnlySpan<byte> Kid => "kid"u8;
     }
-
 }
